+++ repos/omeka-i18n-module/config/services.config.php
@@ -35,9 +35,11 @@ use i18n\Resource\TranslatableResourceManager;
 use i18n\Resource\Writer\JsonResourceWriter;
 use i18n\Resource\Writer\TranslatableResourceWriter;
 use i18n\Site\LocalizationListener;
+use i18n\Site\NavigationTranslator;
 use i18n\Site\SiteListener;
 use i18n\Translator\ContextualTranslator;
 use i18n\Translator\DelegateTranslatorFactory;
+use i18n\Translator\NavigationMvcTranslator;
 use Kevinrob\GuzzleCache\CacheMiddleware;
 use Kevinrob\GuzzleCache\Storage\DoctrineCacheStorage;
 use Kevinrob\GuzzleCache\Strategy\PublicCacheStrategy;
@@ -55,7 +57,19 @@ return [
             'transifex.translations' => TransifexObjectFactory::class,
             'transifex' => TransifexClientFactory::class,
 
-            TransifexProjectListener::class => function ($container) {
+            'Omeka\Site\NavigationTranslator' => function (ContainerInterface $container) {
+                return new NavigationTranslator(
+                    $container->get('Omeka\Site\NavigationLinkManager'),
+                    $container->get(NavigationMvcTranslator::class),
+                    $container->get('ViewHelperManager')->get('Url')
+                );
+            },
+
+            NavigationMvcTranslator::class => function (ContainerInterface $container) {
+                return new NavigationMvcTranslator($container->get(ContextualTranslator::class));
+            },
+
+            TransifexProjectListener::class => function (ContainerInterface $container) {
                 return new TransifexProjectListener(
                     $container->get('transifex.projects'),
                     $container->get('Omeka\Settings'),
@@ -63,7 +77,7 @@ return [
                 );
             },
 
-            TranslatableResourceListener::class => function ($container) {
+            TranslatableResourceListener::class => function (ContainerInterface $container) {
                 return new TranslatableResourceListener(
                     $container->get('Omeka\Logger'),
                     $container->get('Omeka\Settings'),
@@ -73,7 +87,7 @@ return [
                 );
             },
 
-            'i18n.guzzle.client' => function ($container) {
+            'i18n.guzzle.client' => function (ContainerInterface $container) {
                 $stack = HandlerStack::create();
                 $stack->push(
                     new CacheMiddleware(
@@ -82,7 +96,7 @@ return [
                                 new ChainCache(
                                     [
                                         new ArrayCache(),
-                                        new FilesystemCache(OMEKA_PATH.'/files'),
+                                        new FilesystemCache(OMEKA_PATH . '/files'),
                                     ]
                                 )
                             )
@@ -106,31 +120,31 @@ return [
                 );
             },
 
-            ContextualTranslator::class => function ($container) {
+            ContextualTranslator::class => function (ContainerInterface $container) {
                 return new ContextualTranslator($container->get('MvcTranslator'));
             },
 
-            TransifexResourceMessageLoader::class => function ($container) {
+            TransifexResourceMessageLoader::class => function (ContainerInterface $container) {
                 return new TransifexResourceMessageLoader(
                     $container->get('Omeka\Logger'),
                     $container->get('transifex.translations')
                 );
             },
 
-            TransifexCoreMessageLoader::class => function ($container) {
+            TransifexCoreMessageLoader::class => function (ContainerInterface $container) {
                 return new TransifexCoreMessageLoader(
                     $container->get('Omeka\Logger'),
                     $container->get('transifex.translations')
                 );
             },
 
-            TransifexThemeMessageLoader::class => function ($container) {
+            TransifexThemeMessageLoader::class => function (ContainerInterface $container) {
                 return new TransifexThemeMessageLoader(
                     $container->get('Omeka\Logger'), $container->get('transifex.translations')
                 );
             },
 
-            LocalizationListener::class => function ($container) {
+            LocalizationListener::class => function (ContainerInterface $container) {
                 $settings = $container->get('Omeka\Settings');
                 $siteSettings = $container->get('Omeka\Settings\Site');
 
@@ -141,7 +155,7 @@ return [
                 );
             },
 
-            TranslatableResourceExporter::class => function ($container) {
+            TranslatableResourceExporter::class => function (ContainerInterface $container) {
                 return new TransifexResourceExporter(
                     $container->get('Omeka\Logger'),
                     $container->get('transifex.resources'),
@@ -149,7 +163,7 @@ return [
                 );
             },
 
-            TranslatableResourceManager::class => function ($container) {
+            TranslatableResourceManager::class => function (ContainerInterface $container) {
                 return new TranslatableResourceManager($container->get('Omeka\ApiManager'));
             },
 
@@ -157,7 +171,7 @@ return [
                 return new JsonResourceWriter();
             },
 
-            TranslationMissingListener::class => function ($container) {
+            TranslationMissingListener::class => function (ContainerInterface $container) {
                 return new TranslationMissingListener(
                     $container->get('Omeka\Logger'),
                     $container->get('transifex.translations')